Move core_alpha config regression test to JVM-only suite
